requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.19
idna==3.4
PyJWT==2.6.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
typing-extensions==4.5.0
urllib3==2.0.2
